tests: Extract kubevirt-resource helpers from utils
